Refactor files to use kebab-case
